+++ src/Simplex/Matrix.hpp
@@ -22,7 +22,7 @@ namespace Simplex
 
     public:
 
-        Matrix(long m, long n, const std::vector<BigInt>& cells);
+        Matrix(long m, long n, const std::vector<long long>& cells);
 
         long getM() const;
